added profile screen
