+++ lib/constants.ts
@@ -150,27 +150,6 @@ const POSTS_CONTENT = [
     },
     shopURL: "/post/23423",
     post: {
-      imageURL: "https://picsum.photos/800",
-      description:
-        "A short travel package without any hassel. Spend weekens in a very reasonable price.",
-      liked: true,
-      likes: "72",
-      comments: "10",
-      shares: "2",
-    },
-    liked: true,
-  },
-  {
-    profile: {
-      name: "Rudrakaal",
-      followers: "1.2M",
-      following: "1.2M",
-      imageURL: "https://picsum.photos/100/100",
-      distance: "3.2 km",
-      id: "23423",
-    },
-    shopURL: "/post/23423",
-    post: {
       imageURL: "https://picsum.photos/800/400",
       description:
         "A short travel package without any hassel. Spend weekens in a very reasonable price.",
@@ -333,4 +312,40 @@ const LOCAL_MARKET_CONTENT = [
   },
 ];
 
-export { TRENDING_CONTENT, POSTS_CONTENT, LOCAL_MARKET_CONTENT };
+const PROFILE = {
+  name: "Manav Singla",
+  designation: "Brand Ambassador",
+  email: "manav.singh@gmail.com",
+  dob: "29 July",
+  bio: "I admire which is natural. Believe in simplicity.",
+  company: "Ustra",
+  id: "23423",
+  imageURL: "https://picsum.photos/100/100",
+  posts: "12",
+  connects: "2123",
+  connecting: "400",
+};
+
+const PROFILE_POSTS = {
+  posts: [
+    "https://picsum.photos/100/100",
+    "https://picsum.photos/300",
+    "https://picsum.photos/500",
+    "https://picsum.photos/600",
+    "https://picsum.photos/700",
+  ],
+
+  discoveries: [
+    "https://picsum.photos/350",
+    "https://picsum.photos/403",
+    "https://picsum.photos/501",
+  ],
+};
+
+export {
+  TRENDING_CONTENT,
+  POSTS_CONTENT,
+  LOCAL_MARKET_CONTENT,
+  PROFILE_POSTS,
+  PROFILE,
+};
